web-scanner-v2/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
colorama>=0.4.6
urllib3>=2.0.0
pyyaml>=6.0
tqdm>=4.66.0
pyjwt>=2.8.0
cryptography>=41.0.0
lxml>=4.9.0
